refactor wf_extensions
